Moving graph helper functions to utils file
